Refactor: 통일성을 위해 ItemRenderer 분리 및 디버깅 요소 제거
